Paths in linked directories need to maintain the path name passed in originally
So we track the logical path name– the one that was passed into to start with,
this is what goes in the .path attribute, and the physical path name, which
is the version that starts with the realpath of the starting root. This is
what goes into the realpath attribute of Node objects. Link objects are
unchanged.