Make sure Nodes always have `.phantomChildren`
Fixes a bug (for me) which occurred when npm was calling
`Object.keys()` on the `.phantomChildren` property of a node even
though there was no such property.

I don’t know why the property was not set in that particular case
and I couldn’t reproduce it once I got things working using this
patch; it doesn’t seem like something that would hurt, though.