Make sure directories for symlinks always exist
Adds the missing test/fixtures/badlink/node_modules directory.

As the directory would usually have no contents except during testing,
it isn’t present in a normal git checkout of this repository and so
needs to be created manually in order for the tests to work.

Work around that by letting the test file create it.